requirements.txt
# Web Framework
fastapi==0.115.8
uvicorn==0.34.0
python-multipart==0.0.18

# Database
pymongo==4.11.3
motor==3.7.0
elasticsearch==8.17.1

# HTTP Client
httpx==0.24.1
requests==2.32.3

# Date/Time
python-dateutil==2.9.0.post0

# Chinese Text Processing
jieba==0.42.1

# Data Processing
pandas==2.2.3

# LLM Providers
openai==1.3.0
anthropic==0.32.0
zhipuai==2.1.5.20250106

# Utilities
python-dotenv==1.0.1
pydantic==2.10.6
python-dotenv==1.0.1
